feat: XYNE-1315 shared model+thinking menu in the Ask AI sidebar
Extracts ModelThinkingSelector out of AIComposer into a shared component
and replaces the sidebar's old Default-labeled ModelSelector with it, so
both surfaces get the same menu: Recommended row (model name first),
search over the allowed models, and the Thinking flyout. The sidebar's
thinking pick rides the same per-message thinkingLevel the composer
sends.
